requirements.txt
matplotlib==3.3.4
mysql-connector-python==8.0.22
numpy==1.19.3
tabulate==0.8.7